TaskGroup: Remove asynchronous behaviour of `TaskCompletionSource`
On Windows all test cases passed. On Linux, the test case `TaskGroupTest(Direct).AddTasksWithError(Cancel)` would fail often. The root cause was that the method in `WaitAll()` which is only `await m_CompleteSource.Task.ConfigureAwait(false);` would actually wait until both tasks were done, not abort immediately (even though documentation says calling `TrySetCancel()` should set the task to a cancelled state immediately). It apparently did so sometimes, but not all the time.

As this looked like a race condition in `TaskCompletionSource`, change to not continue asynchronously, and this appears to make the test case on Linux with .NET SDK 6.0 reliable.

Tested on Windows with the change and it works.

Tested on Linux Ubuntu 22.04 with .NET 6.0 6.0.419 with the entire test suite 62x and no failure. Given it failed within 1-3 tries, this indicates the problem is solved.

Issue: DOTNET-970